Eliminação de Data-Classes através da criação de classes aninhadas. Apenas eram usadas num único sítio.
