+++ src/main/java/org/jfree/data/xy/YIntervalDataItem.java
@@ -37,6 +37,7 @@
 package org.jfree.data.xy;
 
 import org.jfree.data.ComparableObjectItem;
+import java.io.Serializable;
 
 /**
  * An item representing data in the form (x, y, y-low, y-high).
@@ -108,5 +109,92 @@ public class YIntervalDataItem extends ComparableObjectItem {
             return Double.NaN;
         }
     }
+
+
+    /**
+     * A y-interval.  This class is used internally by the
+     * {@link YIntervalDataItem} class.
+     */
+    private static class YInterval implements Serializable {
+
+        /** The y-value. */
+        private double y;
+
+        /** The lower bound of the y-interval. */
+        private double yLow;
+
+        /** The upper bound of the y-interval. */
+        private double yHigh;
+
+        /**
+         * Creates a new instance of {@code YInterval}.
+         *
+         * @param y  the y-value.
+         * @param yLow  the lower bound of the y-interval.
+         * @param yHigh  the upper bound of the y-interval.
+         */
+        private YInterval(double y, double yLow, double yHigh) {
+            this.y = y;
+            this.yLow = yLow;
+            this.yHigh = yHigh;
+        }
+
+        /**
+         * Returns the y-value.
+         *
+         * @return The y-value.
+         */
+        public double getY() {
+            return this.y;
+        }
+
+        /**
+         * Returns the lower bound of the y-interval.
+         *
+         * @return The lower bound of the y-interval.
+         */
+        public double getYLow() {
+            return this.yLow;
+        }
+
+        /**
+         * Returns the upper bound of the y-interval.
+         *
+         * @return The upper bound of the y-interval.
+         */
+        public double getYHigh() {
+            return this.yHigh;
+        }
+
+        /**
+         * Tests this instance for equality with an arbitrary object.
+         *
+         * @param obj  the object ({@code null} permitted).
+         *
+         * @return A boolean.
+         */
+        @Override
+        public boolean equals(Object obj) {
+            if (obj == this) {
+                return true;
+            }
+            if (!(obj instanceof YInterval)) {
+                return false;
+            }
+            YInterval that = (YInterval) obj;
+            if (this.y != that.y) {
+                return false;
+            }
+            if (this.yLow != that.yLow) {
+                return false;
+            }
+            if (this.yHigh != that.yHigh) {
+                return false;
+            }
+            return true;
+        }
+
+    }
+
 
 }
